+++ crates/ruma-signatures/tests/tests.rs
@@ -11,7 +11,7 @@ fn add_key_to_map(public_key_map: &mut PublicKeyMap, name: &str, pair: &Ed25519K
     let version =
         ServerSigningKeyId::from_parts(SigningKeyAlgorithm::Ed25519, pair.version().into());
 
-    sender_key_map.insert(version.to_string(), encoded_public_key);
+    sender_key_map.insert(version, encoded_public_key);
 }
 
 #[test]